Add RAII primitives for HFONT.
I am a little worried about the inability of Scoped_resource to define
ordering semantics.  It is not enough that objects get released.  They
may need to be unwound in a particular order, which is often the case
for GDI objects.  The ability of a Scoped_resource to be movable allows
reordering to happen if the developer isn't careful.